Move device + device-base.js to src/**/*.test.js style tests
This is a non-functional change; before and after the change, the same
number of tests run (55)
