+++ package.json
@@ -25,7 +25,7 @@
   "scripts": {
     "prepublish": "git submodule update --init && npm run build",
     "version": "npm run docs && git add docs",
-    "test": "mocha test test/integration",
+    "test": "mocha test test/integration './src/**/*.test.js'",
     "test:ci": "npm run lint && npm test && npm run coverage",
     "test:silent": "PARTICLE_NOOP=$(npm run test)",
     "lint": "eslint . --ext .js --ignore-path .gitignore",
